Divided nn.c into multiple files.
