+++ src/V1/neural_net.c
@@ -1,89 +1,7 @@
-#include <stdio.h>
+#include "neural_net.h"
+#include "utils.h"
 #include <stdlib.h>
-#include <math.h>
 #include <time.h>
-
-#define INPUT_SIZE 784
-#define HIDDEN_SIZE 128
-#define OUTPUT_SIZE 10
-#define LEARNING_RATE 0.01
-#define EPOCHS 3
-#define BATCH_SIZE 64
-#define NUM_CLASSES 10  // Digits 0-9
-#define VERBOSE 0     // Set to 1 to enable debugging prints
-
-// Timer function
-double get_time(clock_t start) {
-    return (double)(clock() - start) / CLOCKS_PER_SEC;
-}
-
-// Allocate memory for a matrix
-double** allocateMatrix(int rows, int cols) {
-    if (VERBOSE) printf("Allocating matrix of size %d x %d\n", rows, cols);
-    double** mat = (double**)malloc(rows * sizeof(double*));
-    if (!mat) {
-        if (VERBOSE) printf("Failed to allocate matrix rows\n");
-        exit(1);
-    }
-    
-    for (int i = 0; i < rows; i++) {
-        mat[i] = (double*)malloc(cols * sizeof(double));
-        if (!mat[i]) {
-            if (VERBOSE) printf("Failed to allocate matrix columns for row %d\n", i);
-            exit(1);
-        }
-    }
-    if (VERBOSE) printf("Matrix allocation successful\n");
-    return mat;
-}
-
-// Free allocated matrix memory
-void freeMatrix(double** mat, int rows) {
-    if (VERBOSE) printf("Freeing matrix with %d rows\n", rows);
-    for (int i = 0; i < rows; i++) {
-        free(mat[i]);
-    }
-    free(mat);
-    if (VERBOSE) printf("Matrix freed successfully\n");
-}
-
-// Activation functions
-void relu(double* x, int size) {
-    if (VERBOSE) printf("Applying ReLU to vector of size %d\n", size);
-    for (int i = 0; i < size; i++) {
-        x[i] = (x[i] > 0) ? x[i] : 0;
-    }
-    if (VERBOSE) {
-        printf("ReLU output (first 5 elements): ");
-        for (int i = 0; i < 5 && i < size; i++) printf("%.4f ", x[i]);
-        printf("\n");
-    }
-}
-
-void softmax(double* x, int size) {
-    if (VERBOSE) printf("Applying softmax to vector of size %d\n", size);
-    double sum = 0;
-    for (int i = 0; i < size; i++) {
-        x[i] = exp(x[i]);
-        sum += x[i];
-    }
-    for (int i = 0; i < size; i++) {
-        x[i] /= sum;
-    }
-    if (VERBOSE) {
-        printf("Softmax output: ");
-        for (int i = 0; i < size; i++) printf("%.4f ", x[i]);
-        printf("\n");
-    }
-}
-
-// Neural network structure
-typedef struct {
-    double** W1;
-    double** W2;
-    double* b1;
-    double* b2;
-} NeuralNetwork;
 
 // Initialize neural network
 NeuralNetwork* createNetwork() {
@@ -300,74 +218,6 @@ void evaluate(NeuralNetwork* net, double** images, double** labels, int numImage
     if (VERBOSE) printf("Evaluation completed\n");
 }
 
-// Read MNIST dataset
-double** loadMNISTImages(const char* filename, int numImages) {
-    if (VERBOSE) printf("Loading MNIST images from %s...\n", filename);
-    FILE* file = fopen(filename, "rb");
-    if (!file) {
-        printf("Error opening %s\n", filename);
-        exit(1);
-    }
-    fseek(file, 16, SEEK_SET);
-    double** images = allocateMatrix(numImages, INPUT_SIZE);
-    for (int i = 0; i < numImages; i++) {
-        for (int j = 0; j < INPUT_SIZE; j++) {
-            unsigned char pixel;
-
-            if (fread(&pixel, sizeof(unsigned char), 1, file) != 1) {
-                fprintf(stderr, "Error: Failed to read pixel\n");
-                fclose(file);
-                exit(EXIT_FAILURE);
-            }
-
-            images[i][j] = pixel / 255.0;
-        }
-        if (VERBOSE && i % 10000 == 0) printf("Loaded %d images\n", i);
-    }
-    fclose(file);
-    
-    if (VERBOSE) {
-        printf("First pixel of first image: %.4f\n", images[0][0]);
-        printf("Last pixel of first image: %.4f\n", images[0][INPUT_SIZE-1]);
-        printf("Image loading complete\n");
-    }
-    return images;
-}
-
-double** loadMNISTLabels(const char* filename, int numLabels) {
-    if (VERBOSE) printf("Loading MNIST labels from %s...\n", filename);
-    FILE* file = fopen(filename, "rb");
-    if (!file) {
-        printf("Error opening %s\n", filename);
-        exit(1);
-    }
-    fseek(file, 8, SEEK_SET);
-    double** labels = allocateMatrix(numLabels, OUTPUT_SIZE);
-    for (int i = 0; i < numLabels; i++) {
-        unsigned char label;
-        if (fread(&label, sizeof(unsigned char), 1, file) != 1) {
-            fprintf(stderr, "Error: Failed to read label\n");
-            fclose(file);
-            exit(EXIT_FAILURE);
-        }
-
-        for (int j = 0; j < OUTPUT_SIZE; j++) {
-            labels[i][j] = (j == label) ? 1.0 : 0.0;
-        }
-        
-        if (VERBOSE && i % 10000 == 0) printf("Loaded %d labels\n", i);
-    }
-    fclose(file);
-    
-    if (VERBOSE) {
-        printf("First label: ");
-        for (int j = 0; j < OUTPUT_SIZE; j++) printf("%.1f ", labels[0][j]);
-        printf("\n");
-        printf("Label loading complete\n");
-    }
-    return labels;
-}
-
 // Free network memory
 void freeNetwork(NeuralNetwork* net) {
     if (VERBOSE) printf("Freeing neural network...\n");
@@ -379,28 +229,32 @@ void freeNetwork(NeuralNetwork* net) {
     if (VERBOSE) printf("Neural network freed\n");
 }
 
-// Main function
-int main() {
-    printf("MNIST Neural Network\n\n");
+// Activation functions
+void relu(double* x, int size) {
+    if (VERBOSE) printf("Applying ReLU to vector of size %d\n", size);
+    for (int i = 0; i < size; i++) {
+        x[i] = (x[i] > 0) ? x[i] : 0;
+    }
+    if (VERBOSE) {
+        printf("ReLU output (first 5 elements): ");
+        for (int i = 0; i < 5 && i < size; i++) printf("%.4f ", x[i]);
+        printf("\n");
+    }
+}
 
-    if (VERBOSE) printf("Loading datasets...\n");
-    double** train_images = loadMNISTImages("../../data/train-images.idx3-ubyte", 60000);
-    double** train_labels = loadMNISTLabels("../../data/train-labels.idx1-ubyte", 60000);
-    double** test_images = loadMNISTImages("../../data/t10k-images.idx3-ubyte", 10000);
-    double** test_labels = loadMNISTLabels("../../data/t10k-labels.idx1-ubyte", 10000);
-
-    if (VERBOSE) printf("Creating neural network...\n");
-    NeuralNetwork* net = createNetwork();
-    
-    if (VERBOSE) printf("Starting training...\n");
-    train(net, train_images, train_labels, 60000);
-    
-    if (VERBOSE) printf("Starting evaluation...\n");
-    evaluate(net, test_images, test_labels, 10000);
-
-    if (VERBOSE) printf("Cleaning up...\n");
-    freeNetwork(net);
-    
-    if (VERBOSE) printf("Program completed successfully\n");
-    return 0;
+void softmax(double* x, int size) {
+    if (VERBOSE) printf("Applying softmax to vector of size %d\n", size);
+    double sum = 0;
+    for (int i = 0; i < size; i++) {
+        x[i] = exp(x[i]);
+        sum += x[i];
+    }
+    for (int i = 0; i < size; i++) {
+        x[i] /= sum;
+    }
+    if (VERBOSE) {
+        printf("Softmax output: ");
+        for (int i = 0; i < size; i++) printf("%.4f ", x[i]);
+        printf("\n");
+    }
 }
