+++ src/V1/nn.h
@@ -1,0 +1,18 @@
+#ifndef NN_H
+#define NN_H
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <math.h>
+#include <time.h>
+
+#define INPUT_SIZE 784
+#define HIDDEN_SIZE 128
+#define OUTPUT_SIZE 10
+#define LEARNING_RATE 0.01
+#define EPOCHS 3
+#define BATCH_SIZE 64
+#define NUM_CLASSES 10
+#define VERBOSE 0
+
+#endif
